zavrsene sitnice

+++ routes/web.php
@@ -32,25 +32,27 @@ Route::post('/blog/{id}', [Controller::class, 'storeComment']);
 //users
 Route::get('/login', [UserController::class, 'login'])->name('login');
 Route::post('/authenticate', [UserController::class, 'authenticate']);
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 Route::get('/dashboard', [UserController::class, 'dashboard'])->middleware('auth');
 
 Route::put('/dashboard', [UserController::class, 'editAdmin'])->middleware('auth');
 
 Route::get('/dashboard/reservations', [UserController::class, 'showReservations'])->middleware('auth');
-Route::delete('/dashboard/reservations/{id}', [UserController::class, 'deleteReservation']);
+Route::delete('/dashboard/reservations/{id}', [UserController::class, 'deleteReservation'])->middleware('auth');
 
-Route::get('/dashboard/comments/', [UserController::class, 'showComments']);
-Route::delete('/dashboard/comments/{id}', [UserController::class, 'deleteComment']);
+Route::get('/dashboard/comments/', [UserController::class, 'showComments'])->middleware('auth');
+Route::delete('/dashboard/comments/{id}', [UserController::class, 'deleteComment'])->middleware('auth');
 
 Route::get('/dashboard/addAdmin', [UserController::class, 'showAddAdmin'])->middleware('auth');
 Route::post('/dashboard/addAdmin', [UserController::class, 'storeAdmin'])->middleware('auth');
 
 Route::get('/dashboard/menu', [UserController::class, 'showMenu'])->middleware('auth');
 Route::post('/dashboard/menu', [UserController::class, 'storeMenu'])->middleware('auth');
+Route::delete('/dashboard/menu/{id}', [UserController::class, 'deleteMenu'])->middleware('auth');
 
 Route::get('/dashboard/blog', [UserController::class, 'showBlog'])->middleware('auth');
 Route::post('/dashboard/blog', [UserController::class, 'storeBlog'])->middleware('auth');
+Route::delete('/dashboard/blog/{id}', [UserController::class, 'deleteBlog'])->middleware('auth');
